+++ src/main/java/net/kemitix/slushy/app/archiver/DynamicArchiverProperties.java
@@ -1,24 +1,22 @@
-package net.kemitix.slushy.app.reader;
+package net.kemitix.slushy.app.archiver;
 
 import net.kemitix.slushy.app.DynamicListProcessConfig;
-import net.kemitix.slushy.app.config.DynamicConfig;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 @ApplicationScoped
-public class DynamicReaderConfig
-        implements ReaderProperties, DynamicConfig {
+public class DynamicArchiverProperties
+        implements ArchiverProperties {
 
-    @Inject
-    ReaderConfigMapping configMapping;
+    @Inject ArchiverConfigMapping configMapping;
 
     DynamicListProcessConfig config;
 
     @PostConstruct
     void init() {
-        config = new DynamicListProcessConfig(ReaderConfigMapping.PREFIX, configMapping);
+        config = new DynamicListProcessConfig(ArchiverConfigMapping.PREFIX, configMapping);
     }
 
     @Override
@@ -49,12 +47,6 @@ public class DynamicReaderConfig
     @Override
     public long retryDelay() {
         return config.retryDelay();
-    }
-
-    @Override
-    public int maxSize() {
-        return findIntegerValue(ReaderConfigMapping.PREFIX, MAX_SIZE)
-                .orElseGet(configMapping::maxSize);
     }
 
 }
