+++ src/main/java/net/kemitix/slushy/app/hold/HoldDecisionDueByRoute.java
@@ -15,7 +15,7 @@ public class HoldDecisionDueByRoute
 
     @Inject
     public HoldDecisionDueByRoute(
-            DynamicHoldProperties holdProperties,
+            HoldProperties holdProperties,
             SetDueInDays setDueInDays
     ) {
         this.setDueInDays = setDueInDays;
