+++ src/main/java/net/kemitix/slushy/app/reader/DynamicReaderProperties.java
@@ -1,23 +1,24 @@
-package net.kemitix.slushy.app.archiver;
+package net.kemitix.slushy.app.reader;
 
 import net.kemitix.slushy.app.DynamicListProcessConfig;
+import net.kemitix.slushy.app.config.DynamicConfig;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 @ApplicationScoped
-public class DynamicArchiverConfig
-        implements ArchiverProperties {
+public class DynamicReaderProperties
+        implements ReaderProperties, DynamicConfig {
 
     @Inject
-    ArchiverConfigMapping configMapping;
+    ReaderConfigMapping configMapping;
 
     DynamicListProcessConfig config;
 
     @PostConstruct
     void init() {
-        config = new DynamicListProcessConfig(ArchiverConfigMapping.PREFIX, configMapping);
+        config = new DynamicListProcessConfig(ReaderConfigMapping.PREFIX, configMapping);
     }
 
     @Override
@@ -48,6 +49,12 @@ public class DynamicArchiverConfig
     @Override
     public long retryDelay() {
         return config.retryDelay();
+    }
+
+    @Override
+    public int maxSize() {
+        return findIntegerValue(ReaderConfigMapping.PREFIX, MAX_SIZE)
+                .orElseGet(configMapping::maxSize);
     }
 
 }
