+++ src/main/java/net/kemitix/slushy/app/reject/RejectMoveToTargetListRoute.java
@@ -15,7 +15,7 @@ public class RejectMoveToTargetListRoute
 
     @Inject
     public RejectMoveToTargetListRoute(
-            DynamicRejectProperties rejectProperties,
+            RejectProperties rejectProperties,
             MoveCard moveCard
     ) {
         this.rejectProperties = rejectProperties;
