+++ src/main/java/net/kemitix/slushy/app/status/StatusProperties.java
@@ -6,4 +6,14 @@ public interface StatusProperties {
      * How often to report the status in seconds.
      */
     int logPeriod();
+
+    default String listName() {
+        return System.getenv("SLUSHY_STATUS_LIST");
+    }
+
+    default String cardName() {
+        return System.getenv("SLUSHY_STATUS_CARD");
+    }
+
+
 }
