+++ src/main/java/net/kemitix/slushy/app/trello/queue/AmazonSqsProducers.java
@@ -3,10 +3,8 @@ package net.kemitix.slushy.app.trello.queue;
 import io.quarkus.arc.Unremovable;
 import software.amazon.awssdk.services.sqs.SqsClient;
 
-import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
 
-@ApplicationScoped
 public class AmazonSqsProducers {
 
     @Produces
