+++ src/main/java/net/kemitix/slushy/app/zeroattachment/DynamicZeroAttachmentProperties.java
@@ -6,7 +6,7 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 @ApplicationScoped
-public class DynamicZeroAttachmentConfig
+public class DynamicZeroAttachmentProperties
         implements ZeroAttachmentProperties, DynamicConfig {
 
     @Inject
